+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "inkstats-v2",
-    "version": "1.1.0",
+    "version": "1.1.1",
     "main": "./out/index.js",
     "bin": "./out/index.js",
     "license": "GPL-3.0",
@@ -14,7 +14,10 @@
         "eslint": "^7.20.0",
         "eslint-config-node": "^4.1.0",
         "eslint-plugin-node": "^11.1.0",
+        "jest": "^26.6.3",
         "mocha": "^8.3.0",
+        "supertest": "^6.1.3",
+        "ts-jest": "^26.5.2",
         "ts-node": "^9.1.1",
         "ts-node-dev": "^1.1.2",
         "typescript": "^4.2.2",
@@ -46,7 +49,7 @@
         "build-publish:cleanup": "rm -r out",
         "scripts:git": "npx ts-node scripts/git.ts",
         "testing:build": "yarn build && node . serve -d",
-        "test": "npx ts-node tests/tests.ts"
+        "test": "jest"
     },
     "dependencies": {
         "@types/node": "^14.14.31",
