+++ src/app.ts
@@ -1,0 +1,33 @@
+import * as express from "express";
+
+const app = express()
+const createError = require('http-errors');
+
+const path = require("path");
+
+// Import Routers
+const indexRouter = require('./routes/index')
+
+app.use(express.json());
+app.set("views", path.join(__dirname, "../views"));
+app.set('view engine', 'ejs');
+
+app.use('/', indexRouter);
+
+  // catch 404 and forward to error handler
+  app.use(function(req, res, next) {
+    next(createError(404));
+  });
+  
+  // error handler
+  app.use(function(err, req, res, next) {
+    // set locals, only providing error in development
+    res.locals.message = err.message;
+    res.locals.error = req.app.get('env') === 'development' ? err : {};
+  
+    // render the error page
+    res.status(err.status || 500);
+    res.render(`error`);
+  });
+
+module.exports = app;
