+++ src/index.ts
@@ -1,20 +1,13 @@
 import { port, run_with_color, appName } from "./config";
 import * as express from "express";
+const app = require('./app')
 
 const chalk_init = require('./functions/chalk')
 const sections = require('./functions/command_line_usage')
 
 const commandLineUsage = require('command-line-usage')
 
-const app = express()
-
-const path = require("path");
-
 const logger = require('morgan');
-const createError = require('http-errors');
-
-// Import Routers
-const indexRouter = require('./routes/index')
 
 const CLI_name = appName.toLowerCase()
 
@@ -23,12 +16,6 @@ const yargs = require('yargs')
 const { hideBin } = require('yargs/helpers')
 
 const appport = process.env.PORT || port || 8080
-
-app.use(express.json());
-app.set("views", path.join(__dirname, "../views"));
-app.set('view engine', 'ejs');
-
-app.use('/', indexRouter);
 
 
 const usageargs = process.argv.slice(2)
@@ -86,19 +73,3 @@ yargs(hideBin(process.argv))
   })
 
   .argv
-
-  // catch 404 and forward to error handler
-app.use(function(req, res, next) {
-  next(createError(404));
-});
-
-// error handler
-app.use(function(err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render(`error`);
-});
