+++ tests/Test.test.ts
@@ -1,0 +1,13 @@
+const app = require('../out/app')
+const request = require("supertest")
+
+describe("Express Server Root", () => {
+  test("Stats Code should return 200", done => {
+    request(app)
+      .get("/")
+      .then(response => {
+        expect(response.statusCode).toBe(200);
+        done();
+      });
+  });
+});
