--- tests/tests.ts
@@ -1,3 +1,0 @@
-// Coming Soon!
-
-console.log('Tests Coming Soon :)')
